+++ benches/ray_mesh_intersection.rs
@@ -45,7 +45,7 @@ fn ray_mesh_intersection(c: &mut Criterion) {
 
     for vertices_per_side in [10_u32, 100, 1000] {
         group.bench_function(format!("{}_vertices", vertices_per_side.pow(2)), |b| {
-            let ray = Ray3d::new(Vec3::new(0.0, 1.0, 0.0), Vec3::new(0.0, -1.0, 0.0));
+            let ray = Ray3dExt::new(Vec3::new(0.0, 1.0, 0.0), Vec3::new(0.0, -1.0, 0.0));
             let mesh_to_world = Mat4::IDENTITY;
             let mesh = mesh_creation(vertices_per_side);
 
@@ -69,7 +69,7 @@ fn ray_mesh_intersection_no_intersection(c: &mut Criterion) {
 
     for vertices_per_side in [10_u32, 100, 1000] {
         group.bench_function(format!("{}_vertices", (vertices_per_side).pow(2)), |b| {
-            let ray = Ray3d::new(Vec3::new(0.0, 1.0, 0.0), Vec3::new(1.0, 0.0, 0.0));
+            let ray = Ray3dExt::new(Vec3::new(0.0, 1.0, 0.0), Vec3::new(1.0, 0.0, 0.0));
             let mesh_to_world = Mat4::IDENTITY;
             let mesh = mesh_creation(vertices_per_side);