+++ examples/mouse_picking_deferred.rs
@@ -28,8 +28,11 @@ fn setup(
     commands.spawn(PointLightBundle::default());
     commands.spawn((
         PbrBundle {
-            mesh: meshes.add(Mesh::from(shape::UVSphere::default())),
-            material: materials.add(Color::GRAY.into()),
+            mesh: meshes.add(Mesh::from(Sphere::default())),
+            material: materials.add(StandardMaterial {
+                base_color: Color::GRAY,
+                    ..default()
+            }),
             transform: Transform::from_xyz(0.0, 0.0, -5.0),
             ..default()
         },