+++ examples/simplified_mesh.rs
@@ -36,17 +36,20 @@ fn setup_scene(
     commands.spawn((
         PbrBundle {
             // This is a very complex mesh that will be hard to raycast on
-            mesh: meshes.add(Mesh::from(shape::UVSphere {
+            mesh: meshes.add(Mesh::from(shape::UVSphere {//TODO represent as Sphere as UVSphere is deprecated now
                 radius: 1.0,
                 sectors: 1000,
                 stacks: 1000,
             })),
-            material: materials.add(Color::rgb(1.0, 1.0, 1.0).into()),
+            material: materials.add(StandardMaterial {
+                base_color: Color::rgb(1.0, 1.0, 1.0),
+                ..default()
+            }),
             transform: Transform::from_translation(Vec3::new(0.0, 0.0, -5.0)),
             ..default()
         },
         SimplifiedMesh {
-            mesh: meshes.add(Mesh::from(shape::UVSphere::default())),
+            mesh: meshes.add(Mesh::from(Sphere::default())),
         },
     ));
     commands.spawn(PointLightBundle {
@@ -92,7 +95,7 @@ fn setup_ui(mut commands: Commands) {
                 },
                 ..default()
             })
-            .insert(FpsText);
+                .insert(FpsText);
 
             ui.spawn(TextBundle {
                 text: Text {
@@ -118,7 +121,7 @@ fn setup_ui(mut commands: Commands) {
                 },
                 ..default()
             })
-            .insert(SimplifiedStatus);
+                .insert(SimplifiedStatus);
         });
 }
 
@@ -133,7 +136,7 @@ fn manage_simplified_mesh(
     mut commands: Commands,
     query: Query<(Entity, Option<&SimplifiedMesh>), With<Handle<Mesh>>>,
     mut status_query: Query<&mut Text, With<SimplifiedStatus>>,
-    keyboard: Res<Input<KeyCode>>,
+    keyboard: Res<ButtonInput<KeyCode>>,
     mut meshes: ResMut<Assets<Mesh>>,
 ) {
     if keyboard.just_pressed(KeyCode::Space) {
@@ -141,7 +144,7 @@ fn manage_simplified_mesh(
             if let Ok(mut text) = status_query.get_single_mut() {
                 if simplified_mesh.is_none() {
                     commands.entity(entity).insert(SimplifiedMesh {
-                        mesh: meshes.add(Mesh::from(shape::UVSphere::default())),
+                        mesh: meshes.add(Mesh::from(Sphere::default())),
                     });
                     text.sections[1].value = "ON".to_string();
                     text.sections[1].style.color = Color::GREEN;
@@ -157,7 +160,7 @@ fn manage_simplified_mesh(
 
 fn update_fps(diagnostics: Res<DiagnosticsStore>, mut query: Query<&mut Text, With<FpsText>>) {
     for mut text in &mut query {
-        if let Some(fps) = diagnostics.get(FrameTimeDiagnosticsPlugin::FPS) {
+        if let Some(fps) = diagnostics.get(&FrameTimeDiagnosticsPlugin::FPS) {
             if let Some(average) = fps.average() {
                 // Update the value of the second section
                 text.sections[1].value = format!("{:.2}", average);