+++ CHANGELOG.md
@@ -75,7 +75,7 @@
   method.
 - Removed the `Intersection` component. Intersection data can be found using
   `RaycastMesh::intersections()` and `RaycastSource::intersections()`.
-- Changed: `Ray3d::from_screenspace` start from the near plane.
+- Changed: `Ray3dExt::from_screenspace` start from the near plane.
 - Fixed: Raycasts do not hit bottoms of un-rotated `RayCastMesh`es.
 - Changed: `DefaultPluginState` renamed to `RaycastPluginState`.